\name{functions}
\alias{addMapability}
\alias{balancedCenter}
\alias{bedAnnotationLength}
\alias{binGC}
\alias{binParams}
\alias{binReads}
\alias{breakPointInt}
\alias{calcWindParams}
\alias{chrName}
\alias{combineBins}
\alias{combineGC}
\alias{dividePeaks}
\alias{doBinning}
\alias{doCorrection}
\alias{doMapCorrect}
\alias{fdrRate}
\alias{getChrLength}
\alias{getChrPloidy}
\alias{getNumReads}
\alias{getReadInfo}
\alias{getSegs}
\alias{initRdClass}
\alias{listify}
\alias{loessCorrect}
\alias{logScore}
\alias{makedf}
\alias{makeGcCorrBins}
\alias{methAdjust}
\alias{ploidyPercentages}
\alias{plotLoess}
\alias{plotOverlappingHist}
\alias{plotWindows}
\alias{readEntrypoints}
\alias{readParameters}
\alias{replaceNAsWZeros}
\alias{returnOutliers}
\alias{rpois.od}
\alias{stripOutliers}
\alias{trim}
\alias{tripDist}
\alias{verifyFiles}
\title{
  readDepth Functions
}
\description{
  miscellaneous internal functions used by the readDepth package
}
